requirements.txt
Django==2.2
flake8
djangorestframework==3.11.0
